+++ packages/runes2/index.js
@@ -1,7 +1,0 @@
-"use strict";
-/**
- * Created by user on 2018/3/16/016.
- */
-const runes_1 = require("./runes");
-module.exports = runes_1.runes;
-//# sourceMappingURL=index.js.map