+++ packages/runes2/CHANGELOG.md
@@ -3,6 +3,98 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.1 (2022-08-24)
+
+
+### 🐛　Bug Fixes
+
+* avoid jest snap with unicode title ([657c033](https://github.com/bluelovers/runes/commit/657c0335741dd04708fa7d5520294a4808c6be56))
+* jest snap with unicode title ([9d94131](https://github.com/bluelovers/runes/commit/9d94131078c18cbbf7c8b08c01baac4bad9d8b8c))
+* test ([201c6f2](https://github.com/bluelovers/runes/commit/201c6f2829dd858276a8f7c845455f8da7df31d2))
+
+
+### ✨　Features
+
+* subtree import and update test ([f7fe9ee](https://github.com/bluelovers/runes/commit/f7fe9ee651054b7dadd47cdecfeecba90f6d4477))
+
+
+### 📦　Code Refactoring
+
+* support cjs and esm [#1](https://github.com/bluelovers/runes/issues/1) ([b064b4d](https://github.com/bluelovers/runes/commit/b064b4dca15533c3285d55ecd92abed989e5644b))
+
+
+### 📚　Documentation
+
+* update readme ([77c0f9e](https://github.com/bluelovers/runes/commit/77c0f9e918cf28cdea8fea739d92ceca090151ee))
+
+
+### 🛠　Build System
+
+* build by tsdx ([98c4511](https://github.com/bluelovers/runes/commit/98c451159bc69b611aa6e6ed528019246c8fca0c))
+* **typescript:** importHelpers ([421d732](https://github.com/bluelovers/runes/commit/421d7324aa0ddd787a223db6066c4ec25b0b6bff))
+
+
+### ♻️　Chores
+
+* avoid https://github.com/microsoft/TypeScript/issues/32224 ([ec12f3f](https://github.com/bluelovers/runes/commit/ec12f3f72ad50c2aea8e92f712988f67c898c550))
+* **release:** publish ([cf88391](https://github.com/bluelovers/runes/commit/cf8839128c6d51a9ad4788b90a15b7e4b579bb34))
+* **release:** publish ([b65669c](https://github.com/bluelovers/runes/commit/b65669ce3aeffada179cbe72c70bcf1868cfffef))
+* **release:** publish ([cd65843](https://github.com/bluelovers/runes/commit/cd658435d85af83ec32ddf4685750151f1091aeb))
+* **release:** publish ([3c05889](https://github.com/bluelovers/runes/commit/3c05889a338373b2f51318411f1a62a7ef86bfba))
+* **release:** publish ([4e99280](https://github.com/bluelovers/runes/commit/4e992800a4afd339969a2223ac6f4c4afe5e0736))
+* **release:** publish ([d346b43](https://github.com/bluelovers/runes/commit/d346b435119a51d5084bee7ed22d3804a437bf81))
+* **release:** publish ([732575f](https://github.com/bluelovers/runes/commit/732575fe2ef918900477ac72ac16ec6421745e5b))
+* **release:** publish ([3e7fa8f](https://github.com/bluelovers/runes/commit/3e7fa8fc8d1ee525d6c87a1149ce516ed9822f5f))
+* **release:** publish ([c731c9e](https://github.com/bluelovers/runes/commit/c731c9e574328381e43ac6b2529f2cf9894bcad0))
+
+
+### 🔖　Miscellaneous
+
+* . ([b05205f](https://github.com/bluelovers/runes/commit/b05205f822d76585f46c95a01c293f1f2a729254))
+* . ([d4281b1](https://github.com/bluelovers/runes/commit/d4281b12b05498652b1315f42af951260f17bfb7))
+* . ([38fe990](https://github.com/bluelovers/runes/commit/38fe990e6cc20b86ece34ba8c756ef6da93c3f59))
+* postpublish new version ([5f082f8](https://github.com/bluelovers/runes/commit/5f082f81904253793321646b8436edff48bfbc24))
+* update test ([0ee1dc6](https://github.com/bluelovers/runes/commit/0ee1dc6a495729d261435aaf74c057a9fc57d13c))
+* . ([977f860](https://github.com/bluelovers/runes/commit/977f860818fd3a01be003a0c7c018fb6155e484c))
+* . ([200577c](https://github.com/bluelovers/runes/commit/200577ca075b8798cb73eb68e4ddb3b788284e42))
+* rename ([5fd70c5](https://github.com/bluelovers/runes/commit/5fd70c533caf7f4e71aa7e44f64ca84b246eeae5))
+* to ts ([ced1ed0](https://github.com/bluelovers/runes/commit/ced1ed0c13c2712d682211cccd746a5b5cb47cff))
+* bump version ([c6ab244](https://github.com/bluelovers/runes/commit/c6ab244accc309bca67a0b390d814b0dd66889e2))
+* Diacritical marks ([af90c79](https://github.com/bluelovers/runes/commit/af90c79f483ec05996fa5dd35bad8b7e6872efb9))
+* bump version ([0b2de15](https://github.com/bluelovers/runes/commit/0b2de1562bd6687fd8e33909dc1ec34a92e9baeb))
+* add extended grapheme clusters support. fix #2 ([17753ea](https://github.com/bluelovers/runes/commit/17753ea5f0a3d2d3ecd60c329cc85b82f223ef14)), closes [#2](https://github.com/bluelovers/runes/issues/2)
+* bump version ([b2c70e8](https://github.com/bluelovers/runes/commit/b2c70e8ebf88cb9a678a1fc781e6bcd3d5219f56))
+* fix #1 ([2d5c149](https://github.com/bluelovers/runes/commit/2d5c1490783dd06bdd1129663653ee98bbeecf0e)), closes [#1](https://github.com/bluelovers/runes/issues/1)
+* update tests & dev deps ([65fd914](https://github.com/bluelovers/runes/commit/65fd914ebde7d532ef8af06a64fa55975f37a712))
+* update tests & dev deps ([74d8e5a](https://github.com/bluelovers/runes/commit/74d8e5a21cc4e39be8c8c729126d15fd214356aa))
+* Update readme.md ([28cb4c1](https://github.com/bluelovers/runes/commit/28cb4c1e497f6e217233ff2755cdf27cdff15b6f))
+* update dev deps ([ceba55e](https://github.com/bluelovers/runes/commit/ceba55eeab5ce09f43b6fdc46636b9ecc072d3d5))
+
+
+### BREAKING CHANGES
+
+* no longer supported `const runes = require('runes2')`
+* support cjs and esm #1
+
+
+
+# 0.4.0 "🔖　Miscellaneous" (2016-12-12)
+
+
+### 🔖　Miscellaneous
+
+* add substring method ([f4b383c](https://github.com/bluelovers/runes/commit/f4b383cc6556304be92b0e601ffb97f0f73e2a9f))
+* tune readme ([238aca8](https://github.com/bluelovers/runes/commit/238aca8aa64fc8333ba05ee04843259691da03ee))
+* update readme ([7d67d3e](https://github.com/bluelovers/runes/commit/7d67d3e74a6d7644579708a9a428e9aea3e60be5))
+* bump version 🚀 ([7c3100b](https://github.com/bluelovers/runes/commit/7c3100b74dd7992930b746f102963c23a8316a33))
+* - Migrate to AVA ([1d1846d](https://github.com/bluelovers/runes/commit/1d1846d664a7a97275ded08e6b3e8d85da43341e))
+* bump version 🚀 ([269f82d](https://github.com/bluelovers/runes/commit/269f82db9302217fc06d637f44457df79b806e7f))
+* initial commit ([8013b6e](https://github.com/bluelovers/runes/commit/8013b6e4021a41d6b579d76b3332c87389c5f092))
+
+
+
+
+
 ## [0.4.16](https://github.com/bluelovers/runes/compare/runes2@0.4.15...runes2@0.4.16) (2022-01-17)